[BUGFIX] Do not log on download path mismatch
If download tracking is enabled, the TrackDownload middleware is invoked
for every single page URI. This means that there is a mismatch in the
most cases which led to a warning written to the TYPO3 log file. This
could cause that log file to grow in size quickly.

Avoid this by doing a simpler check for the incoming request path and
skip the log message in this case. Notice that the log message is kept
in case of invalid download paths, e.g. missing HMAC.