[BUGFIX] Skip download tracking without site
If a valid download path is requested without a matching site, the TYPO3
SiteResolver middleware resolves to a NullSite. In this case we cannot
have any Matomo tracking configuration. We now catch this case and skip
the download tracking. We still redirect to the file since the file
would have been accessible anyways if download tracking was not enabled.